refactor: adapt json resume education to data
