+++ src/app/resume-page/education-section/get-education-items.ts
@@ -1,7 +1,7 @@
 import { inject, InjectionToken } from '@angular/core'
 import resume from '@/data/resume.json'
 import { Education } from '../data/education'
-import { ADAPT_JSON_RESUME_EDUCATION } from './adapt-json-resume-education'
+import { ADAPT_JSON_RESUME_EDUCATION } from '../data/adapt-json-resume-education'
 
 export type GetEducationItems = () => readonly Education[]
 export const GET_EDUCATION_ITEMS = new InjectionToken<GetEducationItems>(
